Fix retrieval of online resources
