+++ lib/metadata/common/links.js
@@ -82,7 +82,8 @@ function castLink({name, href, protocol, description}) {
     if (resource.isWfsFeatureType()) {
       return {
         serviceURL: resource.getNormalizedWfsServiceLocation(),
-        typeName: resource.getFeatureTypeName()
+        typeName: resource.getFeatureTypeName(),
+        description
       }
     }
     if (resource.isWmsLayer()) {
