+++ lib/metadata/iso/index.js
@@ -6,7 +6,6 @@ const {cleanKeywords} = require('../common/keywords')
 const {normalizeContacts} = require('../common/contacts')
 const {getLicenseFromLinks, getLicenseFromKeywords, getLicenseFromCatalogs} = require('../common/licenses')
 const {castLink} = require('../common/links')
-const {getFileNameFromHref} = require('../common/util')
 const {getAllContacts} = require('./contacts')
 const {getInspireThemeFromKeywords} = require('./themes')
 const {getTopicCategory} = require('./categories')
@@ -17,7 +16,7 @@ const {getStatus} = require('./statuses')
 const {getUpdateFrequency} = require('./update-frequencies')
 const {getDates} = require('./dates')
 const {getAllKeywords} = require('./keywords')
-const {getAllOnLineResources} = require('./onlineResources')
+const {getAllOnLineResources} = require('./online-resources')
 const {getSpatialResolution} = require('./spatial-resolution')
 const {getLicenseFromConstraints} = require('./licenses')
 
@@ -30,8 +29,6 @@ function convert(originalRecord) {
   record.alternateTitle = get(originalRecord, 'identificationInfo.citation.alternateTitle')
   record.description = get(originalRecord, 'identificationInfo.abstract')
   record.type = get(originalRecord, 'hierarchyLevel')
-
-  const featureTypes = []
 
   const keywords = getAllKeywords(originalRecord)
   record.keywords = cleanKeywords(keywords)
@@ -54,19 +51,10 @@ function convert(originalRecord) {
     record.thumbnails = prepareThumbnails(get(originalRecord, 'identificationInfo.graphicOverview'))
 
     // Links
-    const rawLinks = getAllOnLineResources(originalRecord)
-      .map(resource => ({
-        name: resource.name || getFileNameFromHref(resource.linkage),
-        href: resource.linkage,
-        protocol: resource.protocol
-      }))
-      .map(castLink)
+    const rawLinks = getAllOnLineResources(originalRecord).map(castLink)
 
     record.links = rawLinks.filter(result => result.href)
-
-    rawLinks
-      .filter(result => result.typeName)
-      .forEach(ft => featureTypes.push(ft))
+    record.featureTypes = rawLinks.filter(result => result.typeName)
 
     record.license = getLicenseFromLinks(record.links) ||
       getLicenseFromConstraints(get(originalRecord, 'identificationInfo.resourceConstraints', [])) ||
@@ -91,19 +79,13 @@ function convert(originalRecord) {
       const coupledResources = getCoupledResources(originalRecord)
 
       if (record.serviceURL && coupledResources.length > 0) {
-        coupledResources
-          .map(coupledResource => ({
-            relatedTo: hasha(coupledResource.identifier, {algorithm: 'sha1'}),
-            typeName: coupledResource.scopedName,
-            serviceURL: record.serviceURL
-          }))
-          .forEach(ft => featureTypes.push(ft))
+        record.featureTypes = coupledResources.map(coupledResource => ({
+          relatedTo: hasha(coupledResource.identifier, {algorithm: 'sha1'}),
+          typeName: coupledResource.scopedName,
+          serviceURL: record.serviceURL
+        }))
       }
     }
-  }
-
-  if (featureTypes.length > 0) {
-    record.featureTypes = featureTypes
   }
 
   return record
