+++ lib/metadata/iso/services.js
@@ -1,10 +1,12 @@
 'use strict'
 
 const {get, pick, chain} = require('lodash')
-const {getAllKeywords} = require('./keywords')
+
 const {isPublicURL} = require('../../util/url')
-const {getAllOnLineResources} = require('./onlineResources')
 const {getNormalizedWfsServiceLocation} = require('../common/services')
+
+const {getAllKeywords} = require('./keywords')
+const {getAllOnLineResources} = require('./online-resources')
 
 function isWFSService(record) {
   const title = get(record, 'identificationInfo.citation.title', '').toLowerCase()
@@ -19,20 +21,13 @@ function isWFSService(record) {
 }
 
 function getWFSServiceLocation(record) {
-  const onlineResources = getAllOnLineResources(record).map(resource => ({
-    name: resource.name,
-    href: resource.linkage,
-    protocol: resource.protocol
-  }))
+  const onlineResources = getAllOnLineResources(record)
 
   const candidateResources = chain(onlineResources)
     .map(resource => {
       const {linkage, protocol} = resource
-      if (!linkage) {
-        return
-      }
-      if (!isPublicURL(linkage)) {
-        return
+      if (!linkage || !isPublicURL(linkage)) {
+        return null
       }
 
       const hasWfsInLocation = linkage && linkage.toLowerCase().includes('wfs')
@@ -40,17 +35,15 @@ function getWFSServiceLocation(record) {
       if (hasWfsInLocation || hasWfsInProtocol) {
         return resource
       }
+
+      return null
     })
     .compact()
     .value()
 
-  if (candidateResources.length === 0) {
-    return
-  } if (candidateResources.length > 1) {
-    return
+  if (candidateResources.length === 1) {
+    return getNormalizedWfsServiceLocation(candidateResources[0].linkage)
   }
-
-  return getNormalizedWfsServiceLocation(candidateResources[0].linkage)
 }
 
 function getCoupledResources(record) {
